+++ src/storage-preview/HISTORY.rst
@@ -5,7 +5,9 @@ Release History
 
 1.0.0b6
 +++++++
-* `az storage share`: Support for NFS file shares
+* `az storage share/directory create`: Support for NFS file shares
+* `az storage file upload/update/copy start`: Support `--file-mode`, `--owner`, `-group`, override options for copying
+* `az storage file hard-link create`: Support creating hard-link to an NFS file
 
 1.0.0b5
 +++++++